fix: blink area
